Fixed admin action

+++ api/v0/controllers/adminController.js
@@ -1,13 +1,12 @@
 const Task = require("../models/Task");
 const User = require("../models/User");
-const { createError } = require("../utils/createError");
+const {validationResult} = require("express-validator")
 
 const getAllUser = async(req, res, next)=>{
     //Destruct request query
-    const {username, recent, limit, page} = req.query;
+    let {username, recent, limit, page} = req.query;
+    const startIndex = (page - 1)*limit
     try{
-        //Set pagination constraints
-        const {paginationInfo, startIndex} = await paginateResult(User, parseInt(limit), parseInt(page));
         //Initialize users to hold search result
         let users
         //Search by username
@@ -18,54 +17,62 @@ const getAllUser = async(req, res, next)=>{
         else if(recent){
             users = await User.find().sort({$natural: -1}).select("username email").limit(limit).skip(startIndex);
         }
+        //Get all users
         else{
             users = await User.find().select("username email").limit(limit).skip(startIndex);
         }
         if(!users){
-            const error = createError(400, "User not found");
-            return next(error);
+            return res.status(400).json({error: "User not found"})
         }
         return res.status(200).json({
             message: "Users found",
-            output: users,
-            others: paginationInfo
+            output: users
         });
     }catch(err){
-        return next(err);
+        return res.status(500).json({error: "Error fetching users"});
     }
 }
 
-const revokeUser = async (req, res, next)=>{
+const revokeUser = async (req, res)=>{
+    //Validating user request
+    const errors = validationResult(req)
+    if(!errors.isEmpty()){
+        return res.status(400).json({error: errors.array()})
+    }
     //Destruct request query
     const {username} = req.query
     try{
         //Find and delete user
-        await User.findOneAndDelete({username: username})
-        return res.status(200).json({message: "User access has been revoked"})
+        const user = await User.findOneAndDelete({username: username});
+        const userTask = await Task.deleteMany({username: username});
+        if(!userTask){
+            return res.status(200).json({message: "User has been deleted with no tasks found"});
+        }else{
+            return res.status(200).json({message: "User and user's tasks has been deleted"});
+        }
     }catch(err){
-        const error = createError(500, "User revokation failed")
-        return next(error)
+        return res.status(500).json({error: err.message});
     }
 }
 
-const deleteUserTasks = async(req, res, next)=>{
-
-}
-
-const updatePermission = async (req, res, next)=>{
+const updatePermission = async (req, res)=>{
+    //Validating user request
+    const errors = validateResult(req)
+    if(!errors.isEmpty()){
+        return res.status(400).json({error: errors.array()})
+    }
     //Destruct request
     const {username} = req.query
     const {admin} = req.body
     try{
         //Update user admin status 
-        const user = User.findOneAndUpdate({username: username}, {admin: admin});
+        const user = await User.findOneAndUpdate({username: username}, {admin: admin}, {new: true});
         return res.status(200).json({
             message: "Permission update successful",
             output: user
         });
     }catch(err){
-        const error = createError(500, "User permission update failed");
-        return next(error);
+        return res.status(500).json({error: err.message})
     }
 }
